refactor: rename HTMLMarker for CustomMarker
