+++ packages/react-instantsearch-dom-geo/src/index.js
@@ -1,6 +1,6 @@
 export { default as GoogleMapsLoader } from './GoogleMapsLoader';
 export { default as GeoSearch } from './GeoSearch';
 export { default as Marker } from './Marker';
-export { default as HTMLMarker } from './HTMLMarker';
+export { default as CustomMarker } from './CustomMarker';
 export { default as Redo } from './Redo';
 export { default as Control } from './Control';
